+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add custom readiness probe for GeoGebra applets
 - Improve page export readiness detection and add support for multiple readiness probes
 - Ignore `.github` and `test` directories in Docker image build
+- Dump full app configuration to log on startup if `LOG_LEVEL` is set to `DEBUG`
 
 
 ## Version 1.6.0 (2024-07-29)